backend/requirements/base.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
redis
pydantic-settings
PyJWT
email-validator
